Rename configuration-loader to configuration_loader
